many collision changes
Both the player and enemies now use charactercontroller for collisions and movement. they still have to have colliders for special collisions between each other, like taking damage. The player and enemies now have knockback when hit it can be tweaked in the inspector, detection when the player is hit by enemies is still very janky and needs tweaking. Some rooms have been changed to work with the new collision system